refactor: follow and unfollow to redux
