+++ src/Pages/Following/Following.jsx
@@ -1,20 +1,16 @@
+import { useSelector } from "react-redux";
 import { FollowComp, Footer, Header } from "../../Components";
-import { useAuth } from "../../Context/AuthContext/auth-context";
 
 const Following = () => {
-  const {
-    authState: {
-      user: { following },
-    },
-  } = useAuth();
+  const { user } = useSelector((state) => state.auth);
 
   return (
     <div>
       <Header />
       <div className="max-w-2xl mx-auto rounded-lg mt-4 bg-white px-3 py-5 shadow-md">
         <h2 className="px-2 text-2xl font-bold mb-3">Following</h2>
-        {following.length > 0 ? (
-          following.map((item) => (
+        {user.following.length > 0 ? (
+          user.following.map((item) => (
             <FollowComp
               key={item.id}
               id={item._id}
